backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
python-dotenv>=1.0,<2.0
pydantic>=2.8,<3.0
pydantic-settings>=2.3,<3.0
python-multipart>=0.0.9,<1.0
PyMuPDF>=1.24,<2.0
chromadb>=0.5,<1.0
sentence-transformers>=3.0,<4.0
langchain-text-splitters>=0.2,<1.0
google-genai>=0.3,<1.0
requests>=2.32,<3.0
zhipuai>=2.1,<3.0
